+++ src/app/cli/src/cli_commands.rs
@@ -127,6 +127,7 @@ pub fn get_command(
             } else {
                 Box::new(InitCommand::new(
                     cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     submatches.get_flag("exists-ok"),
                     submatches.get_flag("multi-tenant"),
                 ))