+++ src/app/cli/src/commands/login_command.rs
@@ -159,6 +159,13 @@ impl LoginCommand {
 
 #[async_trait::async_trait(?Send)]
 impl Command for LoginCommand {
+    fn needs_workspace(&self) -> bool {
+        match self.scope {
+            odf_server::AccessTokenStoreScope::Workspace => true,
+            odf_server::AccessTokenStoreScope::User => false,
+        }
+    }
+
     async fn run(&mut self) -> Result<(), CLIError> {
         let odf_server_frontend_url = self.get_server_url();